Do not allow collection delete unless empty

+++ app/controllers/admin/delete_collection_controller.rb
@@ -12,8 +12,12 @@ module Admin
     def destroy
       authorize!
 
-      collection.destroy!
-      render_delete_success
+      if collection.works.empty?
+        collection.destroy!
+        render_delete_success
+      else
+        render_delete_failure
+      end
     end
 
     attr_reader :collection_form
@@ -30,6 +34,16 @@ module Admin
       respond_to do |format|
         format.turbo_stream do
           render turbo_stream: turbo_stream.action(:full_page_redirect, dashboard_path)
+        end
+      end
+    end
+
+    def render_delete_failure
+      flash[:danger] = I18n.t('messages.collection_delete_failed')
+      # This breaks out of the turbo frame.
+      respond_to do |format|
+        format.turbo_stream do
+          render turbo_stream: turbo_stream.action(:full_page_redirect, collection_path(collection.druid))
         end
       end
     end
